+++ src/Scenes/PauseMenuScene/PauseMenuScene.cpp
@@ -23,7 +23,7 @@ static void goGameScene(const Engine::Entity)
     Engine::Entity optionEntity = core->globalEntities.getEntity("options");
     auto &options = Game::CoreData::entityManager->getComponent<Component::OptionComponent>(optionEntity);
 
-    Game::CoreData::camera->setFovy(options.fov);
+    Game::CoreData::camera->setFovy((float) options.fov);
     Game::CoreData::sceneManager->setScene<Game::GameScene>(true, false);
 }
